+++ src/utils.cpp
@@ -14,6 +14,22 @@ double binom_LL_mv(NumericVector p, double k, double n, NumericVector mean, Nume
 }
 
 
+double multinom_LL_mv(NumericVector p_j, LogicalVector z_j, double k, double n, NumericVector p_i, NumericVector mean, NumericMatrix Q, int i, int j) {
+  double nn = 0;
+  for(int jj = 0; jj < p_j.size(); jj++) {
+    if(z_j[jj]) {
+      nn += exp(p_j[jj]);
+    }
+  }
+  double mm = k*p_j[j] - n*log(nn);
+  for(int ii = 0; ii < p_i.size(); ii++) {
+    mm -= 0.5*(1.0 + (i != ii))*(p_i[i] - mean[i])*Q(i, ii)*(p_i[ii] - mean[ii]);
+  }
+  return mm;
+}
+
+
+
 double pois_LL(double L, double k, double mean, double precision) {
   return k*L - exp(L) - 0.5*precision*(L - mean)*(L - mean);
 }